requirements.txt
streamlit>=1.32.0
altair>=5.4.0
pandas>=2.2.0
numpy>=1.26.0
nltk>=3.8.1
spacy>=3.7.0
networkx>=3.1
matplotlib>=3.8.0
seaborn>=0.13.0
wordcloud>=1.9.0
scikit-learn>=1.4.0
webvtt-py>=0.5.0
reportlab>=4.0.0
Pillow>=10.0.0
pytest>=7.4.0
